Added API to work with current page
Some resources now can be bound to the current page. For example,
this works for actions out of the box.

Pages have TTL which is extended automatically when some action get called.

When page is expired, it is removed from the memory and all data,
associated with this page too.